renamed ipc-channels.js to ipc-message.js
