+++ src/react/components/CompetitionPage.js
@@ -13,7 +13,7 @@ import PopupEditTable from "./PopupEditTable";
 // ipc communication
 import IPCService from "../../shared/ipc/ipcRendererService";
 const ipcRenderer = window.electron.ipcRenderer;
-const ipcChannels = require("../../shared/ipc/ipcChannels");
+const ipcChannels = require("../../shared/ipc-messages");
 
 const USE_BROWSER = false;
